refactor(tests): Extract mock helpers to conftest and add status assertions
Address PR review comments:
- Extract MockStreamResponse and MockClientWithDownload to shared conftest.py
  to avoid duplication between test_cli_push_get.py and test_cli_workflow.py
- Fix return type annotations to use union types (MagicMock | httpx.Response)
  for methods that may return either based on URL routing
- Add explicit status code assertions for upload responses in tests to provide
  clearer error messages when uploads fail
